feat: added settings
